apps/api-gateway/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
httpx>=0.27
slowapi>=0.1.9
limits[redis]>=3.13
redis>=5
pydantic>=2.7
PyJWT[crypto]>=2.9
grpcio>=1.80
protobuf>=5.26
